Make API return full post info instead of just ID
The explore API previously just returned the IDs of the relevant posts
and then the `Post` component used to load it through a fetch request.
This created problems because of too much asynchrous reqests that
actually pointed at the same endpoint.
Also, this increased load on the server.
As this data is not too big, it is more
practical to send all of this data at once.

Other changes:

- Minor typo in requirements.txt
- Set z-index for the `create-new-post` FAB