Serve all static assets on "/app"
Previously, there was no way of serving the files placed directly inside
the build directory. Then, we thought of using a build script to
dynamically generate the routes for them, but this would be very
inefficient. So, we tackle the situation by telling webpack to assume
that app is served on "/app" and this path is served as a static files
directory by flask.